Improved placeholder handling and permission.

+++ src/main/kotlin/com/github/shynixn/mctennis/impl/commandexecutor/MCTennisCommandExecutor.kt
@@ -46,7 +46,7 @@ class MCTennisCommandExecutor @Inject constructor(
         label: String,
         args: Array<out String>
     ): Boolean {
-        if (args.size >= 3 && args[0].equals("create", true)) {
+        if (args.size >= 3 && sender.hasPermission(Permission.COMMAND_EDIT.permission) &&  args[0].equals("create", true)) {
             val name = args[1]
             var displayName = args.asList().stream().skip(2).asSequence().joinToString(" ")
 
@@ -58,32 +58,32 @@ class MCTennisCommandExecutor @Inject constructor(
             return true
         }
 
-        if (args.size == 2 && args[0].equals("delete", true)) {
+        if (args.size == 2 && sender.hasPermission(Permission.COMMAND_EDIT.permission) && args[0].equals("delete", true)) {
             val name = args[1]
             deleteArena(sender, name)
             return true
         }
 
-        if (args.size == 1 && args[0].equals("list", true)) {
+        if (args.size == 1 && sender.hasPermission(Permission.COMMAND_EDIT.permission) && args[0].equals("list", true)) {
             listArena(sender)
             return true
         }
 
-        if (sender is Player && args.size == 3 && args[0].equals("inventory", true)) {
+        if (sender is Player && sender.hasPermission(Permission.COMMAND_EDIT.permission) && args.size == 3 && args[0].equals("inventory", true)) {
             val name = args[1]
             val team = args[2]
             setInventory(sender, name, team)
             return true
         }
 
-        if (sender is Player && args.size == 3 && args[0].equals("armor", true)) {
+        if (sender is Player && sender.hasPermission(Permission.COMMAND_EDIT.permission) && args.size == 3 && args[0].equals("armor", true)) {
             val name = args[1]
             val team = args[2]
             setArmor(sender, name, team)
             return true
         }
 
-        if (sender is Player && args.size >= 2 && args[0].equals("join", true)) {
+        if (sender is Player && sender.hasPermission(Permission.COMMAND_PLAYER.permission) && args.size >= 2 && args[0].equals("join", true)) {
             val name = args[1]
             val team = if (args.size > 2) {
                 Team.values().firstOrNull { e -> e.name == args[2] }
@@ -95,12 +95,12 @@ class MCTennisCommandExecutor @Inject constructor(
             return true
         }
 
-        if (sender is Player && args.size == 1 && args[0].equals("leave", true)) {
+        if (sender is Player && sender.hasPermission(Permission.COMMAND_PLAYER.permission) && args.size == 1 && args[0].equals("leave", true)) {
             leaveGame(sender)
             return true
         }
 
-        if (args.size >= 1 && args[0].equals("reload", true)) {
+        if (args.size >= 1 && sender.hasPermission(Permission.COMMAND_EDIT.permission) && args[0].equals("reload", true)) {
             val name = if (args.size >= 2) {
                 args[1]
             } else {
@@ -111,7 +111,11 @@ class MCTennisCommandExecutor @Inject constructor(
             return true
         }
 
-        if (args.size == 1 && args[0].equals("help", true)) {
+        if (args.size == 1 && args[0].equals(
+                "help",
+                true
+            ) && sender.hasPermission(Permission.COMMAND_EDIT.permission)
+        ) {
             sender.sendMessage("---------MCTennis---------")
             sender.sendMessage(ChatColor.GRAY.toString() + "/mctennis create <name> <displayName>")
             sender.sendMessage(ChatColor.GRAY.toString() + "/mctennis delete <name>")
@@ -144,20 +148,22 @@ class MCTennisCommandExecutor @Inject constructor(
         args: Array<out String>
     ): List<String> {
         if (args.size == 1) {
-            if (sender.hasPermission(Permission.EDIT_GAME.permissionString)) {
+            if (sender.hasPermission(Permission.COMMAND_EDIT.permission)) {
                 return arrayListOf("create", "delete", "list", "join", "leave", "reload", "help", "inventory")
             }
 
-            return arrayListOf("join", "leave")
+            if (sender.hasPermission(Permission.COMMAND_PLAYER.permission)) {
+                return arrayListOf("join", "leave")
+            }
         }
 
-        if (args.size == 2 && (args[0].equals("create", true)
+        if (args.size == 2 && sender.hasPermission(Permission.COMMAND_EDIT.permission) && (args[0].equals("create", true)
                     || args[0].equals("delete", true) || args[0].equals("reload", true))
         ) {
             return arenaRepository.getAll().map { e -> e.name }
         }
 
-        if (sender is Player && args.size >= 2 && (args[0].equals("join", true))) {
+        if (sender is Player && sender.hasPermission(Permission.COMMAND_PLAYER.permission) && args.size >= 2 && (args[0].equals("join", true))) {
             return arenaRepository.getAll().map { e -> e.name }
         }
